// types/supabase.ts
export type Database = {
  public: {
    Tables: {
      DNT: {
        Row: {
          bonne_reponse: string
          categorie: string | null
          created_at: string
          id: number
          question: string
          reponses: string[]
        }
        Insert: {
          bonne_reponse: string
          categorie?: string | null
          created_at?: string
          id?: number
          question: string
          reponses: string[]
        }
        Update: {
          bonne_reponse?: string
          categorie?: string | null
          created_at?: string
          id?: number
          question?: string
          reponses?: string[]
        }
        Relationships: []
      }
      joueurs: {
        Row: {
          amis: string[] | null
          created_at: string
          id: string
          partie_quotidienne: boolean | null
          pseudo: string | null
          questionsfaites: number[] | null
        }
        Insert: {
          amis?: string[] | null
          created_at?: string
          id: string
          partie_quotidienne?: boolean | null
          pseudo?: string | null
          questionsfaites?: number[] | null
        }
        Update: {
          amis?: string[] | null
          created_at?: string
          id?: string
          partie_quotidienne?: boolean | null
          pseudo?: string | null
          questionsfaites?: number[] | null
        }
        Relationships: []
      }
      questions: {
        Row: {
          bonne_reponse: string
          categorie: string | null
          created_at: string
          id: number
          question: string
          reponses: string[]
        }
        Insert: {
          bonne_reponse: string
          categorie?: string | null
          created_at?: string
          id?: number
          question: string
          reponses: string[]
        }
        Update: {
          bonne_reponse?: string
          categorie?: string | null
          created_at?: string
          id?: number
          question?: string
          reponses?: string[]
        }
        Relationships: []
      }
      saison_en_cours: {
        Row: {
          bonne_reponse: string
          categorie: string | null
          created_at: string
          id: number
          question: string
          reponses: string[]
        }
        Insert: {
          bonne_reponse: string
          categorie?: string | null
          created_at?: string
          id?: number
          question: string
          reponses: string[]
        }
        Update: {
          bonne_reponse?: string
          categorie?: string | null
          created_at?: string
          id?: number
          question?: string
          reponses?: string[]
        }
        Relationships: []
      }
      saisons_passees: {
        Row: {
          bonne_reponse: string
          categorie: string | null
          created_at: string
          id: number
          question: string
          reponses: string[]
        }
        Insert: {
          bonne_reponse: string
          categorie?: string | null
          created_at?: string
          id?: number
          question: string
          reponses: string[]
        }
        Update: {
          bonne_reponse?: string
          categorie?: string | null
          created_at?: string
          id?: number
          question?: string
          reponses?: string[]
        }
        Relationships: []
      }
      scores: {
        Row: {
          created_at: string
          id: number
          player: string
          score: number
        }
        Insert: {
          created_at?: string
          id?: number
          player: string
          score?: number
        }
        Update: {
          created_at?: string
          id?: number
          player?: string
          score?: number
        }
        Relationships: [
          {
            foreignKeyName: "scores_player_fkey"
            columns: ["player"]
            isOneToOne: true
            referencedRelation: "joueurs"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      get_random_questions: {
        Args: { nb_questions: number; joueur_id: string }
        Returns: {
          id: number
          question: string
          reponse: string
          propositions: string[]
        }[]
      }
      rotate_questions: {
        Args: { count: number }
        Returns: undefined
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never